+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/external/block/controller/BlockServiceController.java
@@ -1,7 +1,7 @@
 package com.blockpage.memberservice.adaptor.infrastructure.external.block.controller;
 
 import com.blockpage.memberservice.adaptor.infrastructure.external.block.requestbody.RequestBlock;
-import com.blockpage.memberservice.application.port.out.BlockPort;
+import com.blockpage.memberservice.application.port.out.port.BlockPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,9 +14,9 @@ public class BlockServiceController implements BlockPort {
     private final BlockServiceOpenFeign blockServiceOpenFeign;
 
     @Override
-    public ResponseEntity postBlock(String email,RequestBlock requestBlock) {
+    public ResponseEntity postBlock(String email, RequestBlock requestBlock) {
         String type = "attendance";
-        blockServiceOpenFeign.postBlock(email,type, requestBlock);
+        blockServiceOpenFeign.postBlock(email, type, requestBlock);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 }
