+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/external/purchase/controller/PurchaseServiceController.java
@@ -1,7 +1,7 @@
 package com.blockpage.memberservice.adaptor.infrastructure.external.purchase.controller;
 
 import com.blockpage.memberservice.adaptor.infrastructure.external.purchase.requestbody.RequestPurchase;
-import com.blockpage.memberservice.application.port.out.PurchasePort;
+import com.blockpage.memberservice.application.port.out.port.PurchasePort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ public class PurchaseServiceController implements PurchasePort {
     @Override
     public ResponseEntity postProfileSkin(String email, RequestPurchase requestPurchase) {
         String type = "profileSkin";
-        purchaseServiceOpenFeign.postProfileSkin(email,type, requestPurchase);
+        purchaseServiceOpenFeign.postProfileSkin(email, type, requestPurchase);
 
         return ResponseEntity.ok().build();
     }
