+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/persistance/AttendanceAdaptor.java
@@ -1,10 +1,10 @@
 package com.blockpage.memberservice.adaptor.infrastructure.persistance;
 
-import static com.blockpage.memberservice.exception.ErrorCode.ATTENDANCE_ALREADY_POST;
+import static com.blockpage.memberservice.exception.ErrorCode.*;
 
 import com.blockpage.memberservice.adaptor.infrastructure.entity.AttendanceEntity;
 import com.blockpage.memberservice.adaptor.infrastructure.repository.AttendanceRepository;
-import com.blockpage.memberservice.application.port.out.AttendancePort;
+import com.blockpage.memberservice.application.port.out.port.AttendancePort;
 import com.blockpage.memberservice.domain.Attendance;
 import com.blockpage.memberservice.exception.CustomException;
 import java.time.LocalDate;
