+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/persistance/EmotionAdaptor.java
@@ -1,10 +1,10 @@
 package com.blockpage.memberservice.adaptor.infrastructure.persistance;
 
-import static com.blockpage.memberservice.exception.ErrorCode.EMOTION_ALREADY_POST;
+import static com.blockpage.memberservice.exception.ErrorCode.*;
 
 import com.blockpage.memberservice.adaptor.infrastructure.entity.EmotionEntity;
 import com.blockpage.memberservice.adaptor.infrastructure.repository.EmotionRepository;
-import com.blockpage.memberservice.application.port.out.EmotionPort;
+import com.blockpage.memberservice.application.port.out.port.EmotionPort;
 import com.blockpage.memberservice.domain.Emotion;
 import com.blockpage.memberservice.exception.CustomException;
 import java.util.List;
