+++ src/main/java/com/blockpage/memberservice/adaptor/web/controller/EmotionController.java
@@ -6,8 +6,8 @@ import com.blockpage.memberservice.application.port.in.EmotionUseCase;
 import com.blockpage.memberservice.application.port.in.EmotionUseCase.DeleteQuery;
 import com.blockpage.memberservice.application.port.in.EmotionUseCase.FindQuery;
 import com.blockpage.memberservice.application.port.in.EmotionUseCase.PostQuery;
-import com.blockpage.memberservice.application.port.in.RequestEmotion;
-import com.blockpage.memberservice.application.port.out.EmotionDto;
+import com.blockpage.memberservice.adaptor.web.requestBody.RequestEmotion;
+import com.blockpage.memberservice.application.port.out.dto.EmotionDto;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
