+++ src/main/java/com/blockpage/memberservice/application/port/in/MemberUseCase.java
@@ -1,6 +1,7 @@
 package com.blockpage.memberservice.application.port.in;
 
-import com.blockpage.memberservice.application.port.out.MemberDto;
+import com.blockpage.memberservice.adaptor.web.requestBody.RequestMember;
+import com.blockpage.memberservice.application.port.out.dto.MemberDto;
 import java.io.IOException;
 import lombok.Builder;
 import lombok.Getter;
@@ -44,7 +45,7 @@ public interface MemberUseCase {
         private String type;
         private String creatorNickname;
 
-        public static FindMemberQuery toQuery(String email,String type, RequestMember requestMember) {
+        public static FindMemberQuery toQuery(String email, String type, RequestMember requestMember) {
             return FindMemberQuery.builder()
                 .email(email)
                 .type(type)
@@ -71,7 +72,7 @@ public interface MemberUseCase {
 
         private MultipartFile profileImage;
 
-        public static UpdateQuery toQuery(String email,String type, RequestMember requestMember, MultipartFile profileImage) {
+        public static UpdateQuery toQuery(String email, String type, RequestMember requestMember, MultipartFile profileImage) {
             return UpdateQuery.builder()
                 .type(type)
                 .email(email)
