+++ src/main/java/com/blockpage/memberservice/application/service/AttendanceService.java
@@ -2,8 +2,8 @@ package com.blockpage.memberservice.application.service;
 
 import com.blockpage.memberservice.adaptor.infrastructure.external.block.requestbody.RequestBlock;
 import com.blockpage.memberservice.application.port.in.AttendanceUseCase;
-import com.blockpage.memberservice.application.port.out.AttendancePort;
-import com.blockpage.memberservice.application.port.out.BlockPort;
+import com.blockpage.memberservice.application.port.out.port.AttendancePort;
+import com.blockpage.memberservice.application.port.out.port.BlockPort;
 import com.blockpage.memberservice.domain.Attendance;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
